Refactored swagger generator
